Refactor: KakaoService Provider 함수 다른 파일로 옮기는 차원에서 삭제
